+++ rts-game/src/app.cpp
@@ -79,7 +79,7 @@ HRESULT App::init(HINSTANCE hInstance, int width, int height) {
     return S_OK;
 }
 
-HRESULT App::update(double dt, double t) {
+HRESULT App::update(float dt, float t) {
     time_elapsed = t;
     return 0;
 }
